Fix workflow concurrency group
The concurrency group was misconfigured because it did not take into
account the matrix values.

The previous version worked like this:

	runs-on: ${{ matrix.os }}

	concurrency:
		# this does not take into account the current os
		group: ${{ github.workflow }}-${{ github.ref }}
		cancel-in-progress: true

The new version adds ${{ matrix.os }} to the group definition,
which makes it unique for each job run in the matrix.

This caused a problem in how concurrency was handled, and a CI fail.